requirements.txt
Flask==0.10.1
Flask-Dance[models]==0.3.1
Flask-SQLAlchemy==2.0
Flask-Script
Flask-SSLify
celery==3.1.17
redis==2.10.3
blinker
requests==2.5.0
bugsnag
gunicorn==19.1.1
psycopg2==2.5.4
iso8601
